config: point platform_url() at api.{domain}, relocate query_rewriter
- platform_url(): return https://api.{base_domain}, not https://app.
  Clients already prepend /v1/... themselves. The app.{domain} default
  was the user-facing web app and required CASTFORM_PLATFORM_URL to be
  overridden in every script.
- query_rewriter: move out of envs/postgres_search/ — it has no postgres
  dependency and its only caller is rag/qa_generation/retrieval_query.
